refactor(ktor): isolate Kotlinx default contract
