+++ client/webauthn-client-ktor-kotlinx/README.md
@@ -5,3 +5,10 @@ Kotlinx Serialization implementation of the default Ktor passkey backend contrac
 Use this only with the default `/webauthn/…` payload shape. Apps that use another JSON library or
 backend contract should depend on `webauthn-client-ktor` and provide their own
 `KtorPasskeyContractCodec` implementation.
+
+## What it provides
+
+- `KotlinxKtorPasskeyBackend` and `KotlinxKtorPasskeyContractCodec` for typed flow backends
+- `KtorPasskeyServerClient`, the legacy default-contract client during controller migration
+- `RegistrationStartPayload` and `AuthenticationStartPayload` for the default request shape
+- `KtorOriginMetadataProvider`, which fails closed on fetch or parse errors
