+++ client/webauthn-client-ktor-kotlinx/build.gradle.kts
@@ -12,11 +12,18 @@ kotlin {
     sourceSets {
         commonMain.dependencies {
             api(project(":client:webauthn-client-ktor"))
+            api(project(":client:webauthn-client-flow"))
+            api(project(":core:webauthn-core"))
             api(project(":core:webauthn-json-kotlinx"))
+            implementation(project(":core:webauthn-runtime-core"))
+            implementation(libs.ktor.client.content.negotiation)
+            implementation(libs.ktor.serialization.kotlinx.json)
             implementation(libs.kotlinx.serialization.json)
         }
         commonTest.dependencies {
             implementation(kotlin("test"))
+            implementation(libs.ktor.client.mock)
+            implementation(libs.kotlinx.coroutines.test)
         }
     }
 }
