+++ client/webauthn-client-ktor/build.gradle.kts
@@ -1,7 +1,6 @@
 plugins {
     id("webauthn.kotlin.multiplatform")
     id("webauthn.published-library")
-    alias(libs.plugins.kotlin.serialization)
 }
 
 kotlin {
@@ -12,12 +11,7 @@ kotlin {
     sourceSets {
         commonMain.dependencies {
             api(project(":client:webauthn-client-flow"))
-            api(project(":core:webauthn-core"))
-            api(project(":core:webauthn-json-kotlinx"))
-            implementation(project(":core:webauthn-runtime-core"))
             api(libs.ktor.client.core)
-            implementation(libs.ktor.client.content.negotiation)
-            implementation(libs.ktor.serialization.kotlinx.json)
         }
         commonTest.dependencies {
             implementation(kotlin("test"))
