+++ documentation/example-inventory.md
@@ -25,9 +25,9 @@ Managed blocks: **107**
 | readme-mermaid-2 | README.md:82 | Repository structure | mermaid | consumer | illustrative | Markdown illustration | illustrative | Diagram is rendered by the Markdown host |
 | readme-kotlin-1 | README.md:146 | Install | kotlin | consumer | configuration | documentation/consumer-smoke/repositories.gradle.kts#consumer-repositories | consumer-compile |  |
 | readme-kotlin-2 | README.md:158 | Install | kotlin | consumer | configuration | documentation/consumer-smoke/client/build.gradle.kts.template#consumer-client-kmp-dependencies | consumer-compile |  |
-| readme-kotlin-3 | README.md:180 | Install | kotlin | consumer | configuration | documentation/consumer-smoke/server/build.gradle.kts.template#consumer-server-dependencies | consumer-compile |  |
-| readme-bash-1 | README.md:200 | Install | bash | maintainer | markdown | Markdown block | syntax |  |
-| readme-bash-2 | README.md:291 | Maintainer Workflow | bash | maintainer | markdown | Markdown block | syntax |  |
+| readme-kotlin-3 | README.md:181 | Install | kotlin | consumer | configuration | documentation/consumer-smoke/server/build.gradle.kts.template#consumer-server-dependencies | consumer-compile |  |
+| readme-bash-1 | README.md:201 | Install | bash | maintainer | markdown | Markdown block | syntax |  |
+| readme-bash-2 | README.md:292 | Maintainer Workflow | bash | maintainer | markdown | Markdown block | syntax |  |
 | client-webauthn-client-compose-readme-mermaid-1 | client/webauthn-client-compose/README.md:13 | What it provides | mermaid | consumer | illustrative | Markdown illustration | illustrative | Diagram is rendered by the Markdown host |
 | client-webauthn-client-compose-readme-kotlin-1 | client/webauthn-client-compose/README.md:34 | How to use | kotlin | consumer | source | documentation/examples/src/platformMain/kotlin/dev/webauthn/documentation/examples/ComposeClientExample.kt#compose-client | platform-compile |  |
 | client-webauthn-client-core-readme-mermaid-1 | client/webauthn-client-flow/README.md:12 | What it provides | mermaid | consumer | illustrative | Markdown illustration | illustrative | Diagram is rendered by the Markdown host |
@@ -35,8 +35,8 @@ Managed blocks: **107**
 | client-webauthn-client-core-readme-kotlin-2 | client/webauthn-client-flow/README.md:94 | Capabilities API | kotlin | consumer | source | documentation/examples/src/commonMain/kotlin/dev/webauthn/documentation/examples/ClientCapabilitiesExample.kt#client-capabilities | compile |  |
 | client-webauthn-client-json-core-readme-kotlin-1 | client/webauthn-client-json-core/README.md:18 | How to use | kotlin | consumer | source | documentation/examples/src/commonMain/kotlin/dev/webauthn/documentation/examples/JsonClientExample.kt#json-client | compile |  |
 | client-webauthn-client-json-core-readme-mermaid-1 | client/webauthn-client-json-core/README.md:34 | How it fits | mermaid | consumer | illustrative | Markdown illustration | illustrative | Diagram is rendered by the Markdown host |
-| client-webauthn-client-ktor-readme-kotlin-1 | client/webauthn-client-ktor/README.md:27 | How to use | kotlin | consumer | source | documentation/examples/src/commonMain/kotlin/dev/webauthn/documentation/examples/NetworkClientExample.kt#network-client | compile |  |
-| client-webauthn-client-ktor-readme-mermaid-1 | client/webauthn-client-ktor/README.md:49 | How it fits | mermaid | consumer | illustrative | Markdown illustration | illustrative | Diagram is rendered by the Markdown host |
+| client-webauthn-client-ktor-readme-kotlin-1 | client/webauthn-client-ktor/README.md:26 | How to use | kotlin | consumer | source | documentation/examples/src/commonMain/kotlin/dev/webauthn/documentation/examples/NetworkClientExample.kt#network-client | compile |  |
+| client-webauthn-client-ktor-readme-mermaid-1 | client/webauthn-client-ktor/README.md:48 | How it fits | mermaid | consumer | illustrative | Markdown illustration | illustrative | Diagram is rendered by the Markdown host |
 | client-webauthn-client-platform-readme-kotlin-1 | client/webauthn-client-platform/README.md:23 | Android | kotlin | consumer | source | documentation/examples/src/androidMain/kotlin/dev/webauthn/documentation/examples/AndroidClientExample.kt#android-client | platform-compile |  |
 | client-webauthn-client-platform-readme-kotlin-2 | client/webauthn-client-platform/README.md:38 | iOS | kotlin | consumer | source | documentation/examples/src/iosMain/kotlin/dev/webauthn/documentation/examples/IosClientExample.kt#ios-client | platform-compile |  |
 | client-webauthn-client-platform-readme-mermaid-1 | client/webauthn-client-platform/README.md:51 | How it fits | mermaid | consumer | illustrative | Markdown illustration | illustrative | Diagram is rendered by the Markdown host |
